+++ src/RepetierServerSharpApi/RepetierClient.Files.cs
@@ -97,6 +97,44 @@ namespace AndreasReitberger.API.Repetier
             }
         }
 
+
+        async Task<RepetierModelList?> GetModelListInfoResponeAsync(string printerName)
+        {
+            IRestApiRequestRespone? result = null;
+            try
+            {
+                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{printerName}";
+                result = await SendRestApiRequestAsync(
+                   requestTargetUri: targetUri,
+                   method: Method.Post,
+                   command: "listModels",
+                   jsonObject: null,
+                   authHeaders: AuthHeaders
+                   )
+                .ConfigureAwait(false);
+                RepetierModelList? list = GetObjectFromJson<RepetierModelList>(result?.Result);
+                await UpdateFreeSpaceAsync().ConfigureAwait(false);
+
+                return list;
+            }
+            catch (JsonException jecx)
+            {
+                OnError(new JsonConvertEventArgs()
+                {
+                    Exception = jecx,
+                    OriginalString = result?.Result,
+                    TargetType = nameof(RepetierModelList),
+                    Message = jecx.Message,
+                });
+                return new RepetierModelList();
+            }
+            catch (Exception exc)
+            {
+                OnError(new UnhandledExceptionEventArgs(exc, false));
+                return new RepetierModelList();
+            }
+        }
+
         public async Task<Dictionary<long, byte[]>?> GetModelImagesAsync(ObservableCollection<IGcode> models, GcodeImageType imageType = GcodeImageType.Thumbnail)
         {
             string currentPrinter = GetActivePrinterSlug();
@@ -263,6 +301,110 @@ namespace AndreasReitberger.API.Repetier
         #endregion
 
         #region Groups
+
+        [Obsolete("Use GetModelGroupsAsync() instead")]
+        async Task<RepetierModelGroups?> GetModelGroupsFromPrinterAsync(string printerName)
+        {
+            IRestApiRequestRespone? result = null;
+            try
+            {
+                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{printerName}";
+                result = await SendRestApiRequestAsync(
+                   requestTargetUri: targetUri,
+                   method: Method.Post,
+                   command: "listModelGroups",
+                   jsonObject: null,
+                   authHeaders: AuthHeaders
+                   )
+                .ConfigureAwait(false);
+                return GetObjectFromJson<RepetierModelGroups>(result?.Result);
+            }
+            catch (JsonException jecx)
+            {
+                OnError(new JsonConvertEventArgs()
+                {
+                    Exception = jecx,
+                    OriginalString = result?.Result,
+                    TargetType = nameof(RepetierModelGroups),
+                    Message = jecx.Message,
+                });
+                return new RepetierModelGroups();
+            }
+            catch (Exception exc)
+            {
+                OnError(new UnhandledExceptionEventArgs(exc, false));
+                return new RepetierModelGroups();
+            }
+        }
+        /// <summary>
+        /// Gets the model groups from the server.
+        /// </summary>
+        /// <param name="path">Printer name</param>
+        /// <returns>Available groups</returns>
+        public override async Task<List<IGcodeGroup>> GetModelGroupsAsync(string path = "")
+        {
+            IRestApiRequestRespone? result = null;
+            List<IGcodeGroup> resultObject = [];
+
+            string currentPrinter = !string.IsNullOrEmpty(path) ? path : GetActivePrinterSlug();
+            if (string.IsNullOrEmpty(currentPrinter)) return resultObject;
+
+            try
+            {
+                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
+                result = await SendRestApiRequestAsync(
+                       requestTargetUri: targetUri,
+                       method: Method.Post,
+                       command: "listModelGroups",
+                       jsonObject: null,
+                       authHeaders: AuthHeaders
+                       )
+                    .ConfigureAwait(false);
+                RepetierModelGroups? info = GetObjectFromJson<RepetierModelGroups>(result?.Result);
+                return
+                    info is not null && info.GroupNames is not null ?
+                    [.. info.GroupNames.Select(g => new RepetierModelGroup() { Name = g })] : resultObject;
+            }
+            catch (JsonException jecx)
+            {
+                OnError(new JsonConvertEventArgs()
+                {
+                    Exception = jecx,
+                    OriginalString = result?.Result,
+                    TargetType = nameof(String),
+                    Message = jecx.Message,
+                });
+                return [];
+            }
+            catch (Exception exc)
+            {
+                OnError(new UnhandledExceptionEventArgs(exc, false));
+                return resultObject;
+            }
+        }
+        public async Task RefreshModelGroupsAsync()
+        {
+            try
+            {
+                List<IGcodeGroup> groups = [];
+                if (!IsReady || ActivePrinter == null)
+                {
+                    Groups = [.. groups];
+                    return;
+                }
+
+                string currentPrinter = ActivePrinter.Slug;
+                if (string.IsNullOrEmpty(currentPrinter)) return;
+
+                groups = await GetModelGroupsAsync(path: currentPrinter).ConfigureAwait(false);
+                Groups = [.. groups];
+            }
+            catch (Exception exc)
+            {
+                OnError(new UnhandledExceptionEventArgs(exc, false));
+                Groups = [];
+            }
+        }
         public async Task<bool> AddModelGroupAsync(IGcodeGroup? group) => await AddModelGroupAsync(group?.Name);
 
         public async Task<bool> AddModelGroupAsync(string? groupName)
@@ -480,78 +622,6 @@ namespace AndreasReitberger.API.Repetier
             }
         }
 
-        public async Task<List<IGcodeGroup>> GetModelGroupsAsync()
-        {
-            IRestApiRequestRespone? result = null;
-            List<IGcodeGroup> resultObject = [];
-
-            string currentPrinter = GetActivePrinterSlug();
-            if (string.IsNullOrEmpty(currentPrinter)) return resultObject;
-
-            try
-            {
-                string targetUri = $"{RepetierCommands.Base}/{RepetierCommands.Api}/{currentPrinter}";
-                result = await SendRestApiRequestAsync(
-                       requestTargetUri: targetUri,
-                       method: Method.Post,
-                       command: "listModelGroups",
-                       jsonObject: null,
-                       authHeaders: AuthHeaders
-                       )
-                    .ConfigureAwait(false);
-                /*
-                result = await SendRestApiRequestAsync(
-                   commandBase: RepetierCommandBase.printer,
-                   commandFeature: RepetierCommandFeature.api, 
-                   command: "listModelGroups", 
-                   printerName: currentPrinter)
-                    .ConfigureAwait(false);
-                */
-                RepetierModelGroups? info = GetObjectFromJson<RepetierModelGroups>(result?.Result);
-                return 
-                    info is not null && info.GroupNames is not null ? 
-                    [.. info.GroupNames.Select(g => new RepetierModelGroup() { Name = g })] : resultObject;
-            }
-            catch (JsonException jecx)
-            {
-                OnError(new JsonConvertEventArgs()
-                {
-                    Exception = jecx,
-                    OriginalString = result?.Result,
-                    TargetType = nameof(String),
-                    Message = jecx.Message,
-                });
-                return [];
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                return resultObject;
-            }
-        }
-        public async Task RefreshModelGroupsAsync()
-        {
-            try
-            {
-                List<IGcodeGroup> groups = [];
-                if (!IsReady || ActivePrinter == null)
-                {
-                    Groups = [.. groups];
-                    return;
-                }
-
-                string currentPrinter = ActivePrinter.Slug;
-                if (string.IsNullOrEmpty(currentPrinter)) return;
-
-                RepetierModelGroups? result = await GetModelGroupsAsync(currentPrinter).ConfigureAwait(false);
-                Groups = result is not null ? ([.. result.GroupNames.Select(g => new RepetierModelGroup() { Name = g })]) : ([.. groups]);
-            }
-            catch (Exception exc)
-            {
-                OnError(new UnhandledExceptionEventArgs(exc, false));
-                Groups = [];
-            }
-        }
         #endregion
 
         #endregion
